requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
torch==2.2.2
torchvision==0.17.2
numpy==1.24.2
Pillow==9.5.0
python-dotenv==1.0.0 